# docs/swagger.yaml
definitions:
...
  request.AssignSave:
    properties:
      api_list:
        items:
          type: integer
        type: array
      id:
        type: integer
      menu_list:
        items:
          type: integer
        type: array
    type: object
...
  request.MenuList:
    properties:
      status:
        type: string
      title:
        type: string
    type: object
...
  request.ReqMenu:
    properties:
      code:
        maxLength: 255
        minLength: 2
        type: string
      component:
        type: string
      hidden:
        type: integer
      icon:
        type: string
      is_link:
        type: integer
      keep_alive:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      sort:
        type: integer
      status:
        type: integer
      title:
        maxLength: 255
        minLength: 2
        type: string
      type:
        type: integer
    required:
    - code
    - sort
    - status
    - title
    type: object
  request.ReqRole:
    properties:
      id:
        type: integer
      parent_id:
        type: integer
      remark:
        type: string
      role_name:
        maxLength: 255
        minLength: 2
        type: string
      status:
        type: integer
    required:
    - role_name
    type: object
...
host: 127.0.0.1:9111
info:
  contact: {}
  description: gin框架后台管理
  title: gin框架后台管理demo
  version: "1.0"
paths:
...
  /api/v1/menu/getMenuList:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 描述
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MenuList'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[],"msg":"获取菜单列表成功！"}'
          schema:
            type: string
      summary: 菜单列表
      tags:
      - Menu
  /api/v1/menu/getSort/:id:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[],"msg":"获得子菜单sort最大值成功！"}'
          schema:
            type: string
      summary: 子菜单sort最大值
      tags:
      - Menu
  /api/v1/menu/menu:
    post:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 菜单名称,权限code,排序
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqMenu'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"添加菜单成功！"}'
          schema:
            type: string
      summary: 添加菜单
      tags:
      - Menu
  /api/v1/menu/menu/:id:
    delete:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"删除菜单成功！"}'
          schema:
            type: string
      summary: 删除菜单
      tags:
      - Menu
    put:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 菜单名称,权限code,排序
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqMenu'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"编辑菜单成功！"}'
          schema:
            type: string
      summary: 编辑菜单
      tags:
      - Menu
  /api/v1/menu/parent:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[],"msg":"获取父级菜单列表成功！"}'
          schema:
            type: string
      summary: 父级菜单列表
      tags:
      - Menu
...
  /api/v1/role/role:
    post:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 角色名称
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqRole'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"添加角色成功！"}'
          schema:
            type: string
      summary: 添加角色
      tags:
      - Role
  /api/v1/role/role/:id:
    delete:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"删除角色成功！"}'
          schema:
            type: string
      summary: 删除角色
      tags:
      - Role
    put:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 角色名称
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqRole'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"编辑角色成功！"}'
          schema:
            type: string
      summary: 编辑角色
      tags:
      - Role
  /api/v1/role/roleAssignSave:
    post:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 角色id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssignSave'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"分配权限成功！"}'
          schema:
            type: string
      summary: 分配权限
      tags:
      - Role
...
swagger: "2.0"